43-refine-rockets-2-3: Restructure `setup.py`
- Divide launch site, motors, and rockets into separate files
- Begin tracking file used for Ascent Test 1 trajectory planning
